# backend_golang/docs/swagger.yaml
basePath: /api/v1
definitions:
  ent.Appointment:
    properties:
      DoctorId:
        description: DoctorId holds the value of the "DoctorId" field.
        type: integer
      UserId:
        description: UserId holds the value of the "UserId" field.
        type: integer
      detail:
        description: Detail holds the value of the "detail" field.
        type: string
      edges:
        $ref: '#/definitions/ent.AppointmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the AppointmentQuery when eager-loading is set.
      endTime:
        description: EndTime holds the value of the "endTime" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      reasonForAppointment:
        description: ReasonForAppointment holds the value of the "reasonForAppointment"
          field.
        type: string
      startTime:
        description: StartTime holds the value of the "startTime" field.
        type: string
      status:
        description: Status holds the value of the "status" field.
        type: string
    type: object
  ent.AppointmentEdges:
    properties:
      appointment_chat:
        $ref: '#/definitions/ent.Chat'
        description: AppointmentChat holds the value of the appointment_chat edge.
      appointment_schedule:
        $ref: '#/definitions/ent.Schedule'
        description: AppointmentSchedule holds the value of the appointment_schedule
          edge.
    type: object
  ent.Certification:
    properties:
      Issuer:
        description: Issuer holds the value of the "Issuer" field.
        type: string
      code:
        description: Code holds the value of the "code" field.
        type: string
      dateOfExp:
        description: DateOfExp holds the value of the "dateOfExp" field.
        type: string
      dateOfIssuing:
        description: DateOfIssuing holds the value of the "dateOfIssuing" field.
        type: string
      diloma:
        description: Diloma holds the value of the "diloma" field.
        type: string
      edges:
        $ref: '#/definitions/ent.CertificationEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CertificationQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.CertificationEdges:
    properties:
      doctor_owner:
        $ref: '#/definitions/ent.User'
        description: DoctorOwner holds the value of the doctor_owner edge.
    type: object
  ent.Chat:
    properties:
      Chat_room_name:
        description: ChatRoomName holds the value of the "Chat_room_name" field.
        type: string
      IsLockChat:
        description: IsLockChat holds the value of the "IsLockChat" field.
        type: boolean
      edges:
        $ref: '#/definitions/ent.ChatEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ChatQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ChatEdges:
    properties:
      chat_appointment:
        description: ChatAppointment holds the value of the chat_appointment edge.
        items:
          $ref: '#/definitions/ent.Appointment'
        type: array
      chat_message:
        description: ChatMessage holds the value of the chat_message edge.
        items:
          $ref: '#/definitions/ent.Message'
        type: array
      chat_user:
        description: ChatUser holds the value of the chat_user edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.Department:
    properties:
      edges:
        $ref: '#/definitions/ent.DepartmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DepartmentQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      image:
        description: Image holds the value of the "image" field.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.DepartmentEdges:
    properties:
      department_has_doctor:
        description: DepartmentHasDoctor holds the value of the department_has_doctor
          edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.Disease:
    properties:
      edges:
        $ref: '#/definitions/ent.DiseaseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DiseaseQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      symtoms:
        description: Symtoms holds the value of the "symtoms" field.
        type: string
    type: object
  ent.DiseaseEdges:
    properties:
      disease_user:
        description: DiseaseUser holds the value of the disease_user edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.Hospital:
    properties:
      edges:
        $ref: '#/definitions/ent.HospitalEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the HospitalQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.HospitalEdges:
    properties:
      hospital_has_doctor:
        description: HospitalHasDoctor holds the value of the hospital_has_doctor
          edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.Message:
    properties:
      edges:
        $ref: '#/definitions/ent.MessageEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MessageQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      message_text:
        description: MessageText holds the value of the "message_text" field.
        type: string
      sent_dateTime:
        description: SentDateTime holds the value of the "sent_dateTime" field.
        type: string
    type: object
  ent.MessageEdges:
    properties:
      What_messages_are_in_this_chat:
        $ref: '#/definitions/ent.Chat'
        description: WhatMessagesAreInThisChat holds the value of the What_messages_are_in_this_chat
          edge.
      Who_send_messages:
        $ref: '#/definitions/ent.User'
        description: WhoSendMessages holds the value of the Who_send_messages edge.
    type: object
  ent.Notification:
    properties:
      AppointmentId:
        description: AppointmentId holds the value of the "AppointmentId" field.
        type: integer
      CreatedDate:
        description: CreatedDate holds the value of the "CreatedDate" field.
        type: string
      Message:
        description: Message holds the value of the "Message" field.
        type: string
      RecipientId:
        description: RecipientId holds the value of the "RecipientId" field.
        type: integer
      SenderId:
        description: SenderId holds the value of the "SenderId" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.NotificationEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the NotificationQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.NotificationEdges:
    properties:
      user_notification:
        description: UserNotification holds the value of the user_notification edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.PInfo:
    properties:
      address:
        description: Address holds the value of the "address" field.
        type: string
      bloodGroup:
        description: BloodGroup holds the value of the "bloodGroup" field.
        type: string
      brithDate:
        description: BrithDate holds the value of the "brithDate" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PInfoEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PInfoQuery when eager-loading is set.
      firstName:
        description: FirstName holds the value of the "firstName" field.
        type: string
      gender:
        description: Gender holds the value of the "gender" field.
        type: integer
      id:
        description: ID of the ent.
        type: integer
      idCardNumber:
        description: IdCardNumber holds the value of the "idCardNumber" field.
        type: string
      lastName:
        description: LastName holds the value of the "lastName" field.
        type: string
    type: object
  ent.PInfoEdges:
    properties:
      who_is_the_owner_of_this_PInfo:
        $ref: '#/definitions/ent.User'
        description: WhoIsTheOwnerOfThisPInfo holds the value of the who_is_the_owner_of_this_PInfo
          edge.
    type: object
  ent.Role:
    properties:
      edges:
        $ref: '#/definitions/ent.RoleEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the RoleQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.RoleEdges:
    properties:
      role_user:
        description: RoleUser holds the value of the role_user edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.Schedule:
    properties:
      activity:
        description: Activity holds the value of the "activity" field.
        type: string
      detail:
        description: Detail holds the value of the "detail" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ScheduleEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ScheduleQuery when eager-loading is set.
      endTime:
        description: EndTime holds the value of the "endTime" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      startTime:
        description: StartTime holds the value of the "startTime" field.
        type: string
      status:
        description: Status holds the value of the "status" field.
        type: string
    type: object
  ent.ScheduleEdges:
    properties:
      schedule_appointment:
        description: ScheduleAppointment holds the value of the schedule_appointment
          edge.
        items:
          $ref: '#/definitions/ent.Appointment'
        type: array
      schedule_doctor:
        $ref: '#/definitions/ent.User'
        description: ScheduleDoctor holds the value of the schedule_doctor edge.
    type: object
  ent.Telecom:
    properties:
      edges:
        $ref: '#/definitions/ent.TelecomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TelecomQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      platform:
        description: Platform holds the value of the "platform" field.
        type: string
      telephone:
        description: Telephone holds the value of the "telephone" field.
        type: string
      username:
        description: Username holds the value of the "username" field.
        type: string
    type: object
  ent.TelecomEdges:
    properties:
      who_is_the_owner_of_this_telecom:
        $ref: '#/definitions/ent.User'
        description: WhoIsTheOwnerOfThisTelecom holds the value of the who_is_the_owner_of_this_telecom
          edge.
    type: object
  ent.Token:
    properties:
      AuthToken:
        description: AuthToken holds the value of the "AuthToken" field.
        type: string
      ExpiresAt:
        description: ExpiresAt holds the value of the "ExpiresAt" field.
        type: string
      GeneratedAt:
        description: GeneratedAt holds the value of the "GeneratedAt" field.
        type: string
      edges:
        $ref: '#/definitions/ent.TokenEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TokenQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.TokenEdges:
    properties:
      authentication_token:
        $ref: '#/definitions/ent.User'
        description: AuthenticationToken holds the value of the authentication_token
          edge.
    type: object
  ent.Treatment:
    properties:
      dateTime:
        description: DateTime holds the value of the "dateTime" field.
        type: string
      edges:
        $ref: '#/definitions/ent.TreatmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TreatmentQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      takeTime:
        description: TakeTime holds the value of the "takeTime" field.
        type: number
      treatmentRecord:
        description: TreatmentRecord holds the value of the "treatmentRecord" field.
        type: string
    type: object
  ent.TreatmentEdges:
    properties:
      treatment_was_recorded_by_doctor:
        $ref: '#/definitions/ent.User'
        description: TreatmentWasRecordedByDoctor holds the value of the treatment_was_recorded_by_doctor
          edge.
      user_is_the_treatment_of_record:
        $ref: '#/definitions/ent.User'
        description: UserIsTheTreatmentOfRecord holds the value of the user_is_the_treatment_of_record
          edge.
    type: object
  ent.User:
    properties:
      edges:
        $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      password:
        description: Password holds the value of the "password" field.
        type: string
      username:
        description: Username holds the value of the "username" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      doctor_has_certification:
        description: DoctorHasCertification holds the value of the doctor_has_certification
          edge.
        items:
          $ref: '#/definitions/ent.Certification'
        type: array
      doctor_record_treatment:
        description: DoctorRecordTreatment holds the value of the doctor_record_treatment
          edge.
        items:
          $ref: '#/definitions/ent.Treatment'
        type: array
      doctor_schedule:
        description: DoctorSchedule holds the value of the doctor_schedule edge.
        items:
          $ref: '#/definitions/ent.Schedule'
        type: array
      from_hospital:
        $ref: '#/definitions/ent.Hospital'
        description: FromHospital holds the value of the from_hospital edge.
      has_department:
        $ref: '#/definitions/ent.Department'
        description: HasDepartment holds the value of the has_department edge.
      user_has_PInfo:
        description: UserHasPInfo holds the value of the user_has_PInfo edge.
        items:
          $ref: '#/definitions/ent.PInfo'
        type: array
      user_have_disease:
        description: UserHaveDisease holds the value of the user_have_disease edge.
        items:
          $ref: '#/definitions/ent.Disease'
        type: array
      user_have_role:
        description: UserHaveRole holds the value of the user_have_role edge.
        items:
          $ref: '#/definitions/ent.Role'
        type: array
      user_have_telecoms:
        description: UserHaveTelecoms holds the value of the user_have_telecoms edge.
        items:
          $ref: '#/definitions/ent.Telecom'
        type: array
      user_have_token:
        description: UserHaveToken holds the value of the user_have_token edge.
        items:
          $ref: '#/definitions/ent.Token'
        type: array
      user_have_treatment:
        description: UserHaveTreatment holds the value of the user_have_treatment
          edge.
        items:
          $ref: '#/definitions/ent.Treatment'
        type: array
      user_send_message:
        description: UserSendMessage holds the value of the user_send_message edge.
        items:
          $ref: '#/definitions/ent.Message'
        type: array
      user_send_notification:
        description: UserSendNotification holds the value of the user_send_notification
          edge.
        items:
          $ref: '#/definitions/ent.Notification'
        type: array
      who_is_in_this_chat:
        description: WhoIsInThisChat holds the value of the who_is_in_this_chat edge.
        items:
          $ref: '#/definitions/ent.Chat'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT FINAL PROJECT Example API
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Create user
      operationId: create-user
      parameters:
      - description: User entity
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/ent.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create user
  /users/{id}:
    delete:
      description: get user by ID
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a user entity by ID
    get:
      description: get user by ID
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user entity by ID
    put:
      consumes:
      - application/json
      description: update user by ID
      operationId: update-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User entity
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/ent.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a user entity by ID
  /users/{username}:
    post:
      description: get user by Username
      parameters:
      - description: User Username
        in: path
        name: username
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a user entity by Username
  /users/getDepartment/{id}:
    get:
      description: list user entities
      operationId: list-user
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List user entities
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"